Updated Windows images for Vulkan
Vulkan can now pull the sdk and runtime as 'latest'. This means that the
latest_version script is no longer required for them, and can be pulled
into the image later.

Newer Vulkan SDKs no longer include the runtime binaries, as these are
usually added to systems via graphics drivers. Rather, it's a separate
set provided from the SDK, and has also been added and pulled as
'latest'.