Added temporary Windows Vulkan workaround
The current VulkanSDK available on Windows, v1.2.182 unfortunately has
an issue where a VkStructureType was misspelled, with a subpass with
three S's.

This unfortunately breaks with the versions after it which do not have
this issue, making it fundamentally incompatible with many libraries
expecting the corrected output.

This effectively rolls back Vulkan to 1.2.176 for the time being.